Moving UIContext out of public package.
