+++ java/java.lsp.server/src/org/netbeans/modules/java/lsp/server/ui/AbstractLspStatusDisplayer.java
@@ -21,6 +21,7 @@ package org.netbeans.modules.java.lsp.server.ui;
 import javax.swing.event.ChangeListener;
 import org.eclipse.lsp4j.MessageType;
 import org.netbeans.modules.java.lsp.server.protocol.ShowStatusMessageParams;
+import org.netbeans.modules.java.lsp.server.protocol.UIContext;
 import org.openide.awt.StatusDisplayer;
 
 public abstract class AbstractLspStatusDisplayer extends StatusDisplayer {
