+++ java/java.lsp.server/test/unit/src/org/netbeans/modules/java/lsp/server/ui/AbstractDialogDisplayerTest.java
@@ -43,6 +43,7 @@ import org.netbeans.modules.java.lsp.server.input.ShowInputBoxParams;
 import org.netbeans.modules.java.lsp.server.input.ShowMutliStepInputParams;
 import org.netbeans.modules.java.lsp.server.input.ShowQuickPickParams;
 import org.netbeans.modules.java.lsp.server.protocol.ShowStatusMessageParams;
+import org.netbeans.modules.java.lsp.server.protocol.UIContext;
 import org.openide.NotifyDescriptor;
 import org.openide.awt.StatusDisplayer;
 import org.openide.util.RequestProcessor;
@@ -59,40 +60,40 @@ public class AbstractDialogDisplayerTest extends NbTestCase {
     
     private static class MockUIContext extends UIContext {
         @Override
-        protected boolean isValid() {
+        public boolean isValid() {
             return true;
         }
 
         @Override
-        protected void showMessage(MessageParams msg) {
+        public void showMessage(MessageParams msg) {
         }
 
         @Override
-        protected CompletableFuture<MessageActionItem> showMessageRequest(ShowMessageRequestParams msg) {
+        public CompletableFuture<MessageActionItem> showMessageRequest(ShowMessageRequestParams msg) {
             return CompletableFuture.completedFuture(null);
         }
 
         @Override
-        protected CompletableFuture<String> showInputBox(ShowInputBoxParams params) {
+        public CompletableFuture<String> showInputBox(ShowInputBoxParams params) {
             return CompletableFuture.completedFuture(null);
         }
 
         @Override
-        protected CompletableFuture<List<QuickPickItem>> showQuickPick(ShowQuickPickParams params) {
+        public CompletableFuture<List<QuickPickItem>> showQuickPick(ShowQuickPickParams params) {
             return CompletableFuture.completedFuture(Collections.emptyList());
         }
 
         @Override
-        protected CompletableFuture<Map<String, Either<List<QuickPickItem>, String>>> showMultiStepInput(ShowMutliStepInputParams params) {
+        public CompletableFuture<Map<String, Either<List<QuickPickItem>, String>>> showMultiStepInput(ShowMutliStepInputParams params) {
             return CompletableFuture.completedFuture(Collections.emptyMap());
         }
 
         @Override
-        protected void logMessage(MessageParams msg) {
+        public void logMessage(MessageParams msg) {
         }
 
         @Override
-        protected StatusDisplayer.Message showStatusMessage(ShowStatusMessageParams msg) {
+        public StatusDisplayer.Message showStatusMessage(ShowStatusMessageParams msg) {
             return null;
         }
     }
